requirements.txt
fastapi
uvicorn[standard]
httpx
openai
python-dotenv
langchain
tiktoken
